refactor: add classes as separate files + adjust filtering logic
Resolves: none.
