+++ Currency/Modules/Currency Details/Presentation/View/Cell/TableViewCell.swift
@@ -1,0 +1,18 @@
+//
+//  TableViewCell.swift
+//  Currency
+//
+//  Created by Loay Ashraf on 28/05/2023.
+//
+
+import UIKit
+
+class TableViewCell: UITableViewCell {
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: .subtitle, reuseIdentifier: reuseIdentifier)
+        selectionStyle = .none
+    }
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
