+++ Currency/Modules/Currency Details/Presentation/View/Storyboard/CurrencyDetails.storyboard
@@ -17,14 +17,22 @@
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <stackView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="nRL-Ui-wUA">
-                                <rect key="frame" x="20" y="79" width="353" height="719"/>
+                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" distribution="equalSpacing" alignment="center" spacing="20" translatesAutoresizingMaskIntoConstraints="NO" id="bFS-Lk-0rj">
+                                <rect key="frame" x="20" y="79" width="353" height="753"/>
                                 <subviews>
-                                    <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="bD7-Yw-xks">
-                                        <rect key="frame" x="0.0" y="0.0" width="353" height="719"/>
-                                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                                    </tableView>
+                                    <stackView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="nRL-Ui-wUA">
+                                        <rect key="frame" x="0.0" y="0.0" width="353" height="753"/>
+                                        <subviews>
+                                            <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="bD7-Yw-xks">
+                                                <rect key="frame" x="0.0" y="0.0" width="353" height="753"/>
+                                                <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                                            </tableView>
+                                        </subviews>
+                                    </stackView>
                                 </subviews>
+                                <constraints>
+                                    <constraint firstItem="nRL-Ui-wUA" firstAttribute="leading" secondItem="bFS-Lk-0rj" secondAttribute="leading" id="FYs-08-Zdz"/>
+                                </constraints>
                             </stackView>
                             <activityIndicatorView opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" style="medium" translatesAutoresizingMaskIntoConstraints="NO" id="vGk-3f-OEe">
                                 <rect key="frame" x="186.66666666666666" y="416" width="20" height="20"/>
@@ -34,11 +42,11 @@
                         <viewLayoutGuide key="safeArea" id="vDu-zF-Fre"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
-                            <constraint firstItem="nRL-Ui-wUA" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" constant="20" id="1qB-ad-bk1"/>
-                            <constraint firstItem="vDu-zF-Fre" firstAttribute="bottom" secondItem="nRL-Ui-wUA" secondAttribute="bottom" constant="20" id="S2D-gs-TbY"/>
+                            <constraint firstItem="bFS-Lk-0rj" firstAttribute="leading" secondItem="vDu-zF-Fre" secondAttribute="leading" constant="20" id="Rp5-xZ-KFw"/>
+                            <constraint firstAttribute="trailing" secondItem="bFS-Lk-0rj" secondAttribute="trailing" constant="20" id="ZmL-fb-zkE"/>
                             <constraint firstItem="vGk-3f-OEe" firstAttribute="centerY" secondItem="5EZ-qb-Rvc" secondAttribute="centerY" id="bSJ-qz-9tC"/>
-                            <constraint firstItem="vDu-zF-Fre" firstAttribute="trailing" secondItem="nRL-Ui-wUA" secondAttribute="trailing" constant="20" id="tUe-cH-Ohy"/>
-                            <constraint firstItem="nRL-Ui-wUA" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" constant="20" id="wyd-9Q-FFN"/>
+                            <constraint firstAttribute="bottom" secondItem="bFS-Lk-0rj" secondAttribute="bottom" constant="20" id="djS-a1-c5b"/>
+                            <constraint firstItem="bFS-Lk-0rj" firstAttribute="top" secondItem="vDu-zF-Fre" secondAttribute="top" constant="20" id="mjy-mD-hV3"/>
                             <constraint firstItem="vGk-3f-OEe" firstAttribute="centerX" secondItem="5EZ-qb-Rvc" secondAttribute="centerX" id="y12-iV-XO7"/>
                         </constraints>
                     </view>
